requirements.txt
# Core ML & NLP
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
accelerate>=0.20.0
peft>=0.4.0
bitsandbytes>=0.39.0

# Data Processing
pandas>=1.5.0
numpy>=1.24.0
pyyaml>=6.0

# IOC Extraction
iocextract>=1.16.0
python-magic>=0.4.27
pdfplumber>=0.9.0
beautifulsoup4>=4.12.0
requests>=2.28.0

# Development
black>=23.0.0
flake8>=6.0.0
pytest>=7.3.0
jupyter>=1.0.0

# Model Serving
fastapi>=0.95.0
uvicorn>=0.21.0
pydantic>=1.10.0
